add s to modules
