+++ src/data_gradients/config/data/caching_utils.py
@@ -3,7 +3,7 @@ from typing import Optional, Any, Union, Callable
 
 import torch
 
-from data_gradients.dataset_adapter.output_mapper.tensor_extractor import NestedDataLookup
+from data_gradients.dataset_adapters.output_mapper.tensor_extractor import NestedDataLookup
 from data_gradients.config.data.typing import SupportedDataType
 from data_gradients.utils.detection import XYXYConverter
