+++ src/data_gradients/dataset_adapters/base_adapter.py
@@ -6,8 +6,8 @@ import torch
 from data_gradients.config.data.typing import SupportedDataType
 from data_gradients.config.data.data_config import DataConfig
 
-from data_gradients.dataset_adapter.formatters.base import BatchFormatter
-from data_gradients.dataset_adapter.output_mapper.dataset_output_mapper import DatasetOutputMapper
+from data_gradients.dataset_adapters.formatters.base import BatchFormatter
+from data_gradients.dataset_adapters.output_mapper.dataset_output_mapper import DatasetOutputMapper
 
 
 class BaseDatasetAdapter(ABC):
