+++ src/data_gradients/dataset_adapters/output_mapper/tensor_extractor.py
