+++ src/data_gradients/managers/abstract_manager.py
@@ -14,7 +14,7 @@ from data_gradients.utils.utils import print_in_box
 from data_gradients.visualize.seaborn_renderer import SeabornRenderer
 from data_gradients.utils.pdf_writer import ResultsContainer, Section, FeatureSummary
 from data_gradients.utils.summary_writer import SummaryWriter
-from data_gradients.sample_iterable.base import BaseSampleIterable
+from data_gradients.sample_iterables.base import BaseSampleIterable
 
 logging.basicConfig(level=logging.INFO)
