+++ src/data_gradients/managers/detection_manager.py
@@ -6,8 +6,8 @@ from data_gradients.config.data.typing import SupportedDataType, FeatureExtracto
 from data_gradients.config.utils import get_grouped_feature_extractors
 from data_gradients.managers.abstract_manager import AnalysisManagerAbstract
 from data_gradients.utils.summary_writer import SummaryWriter
-from data_gradients.dataset_adapter.detection_adapter import DetectionDatasetAdapter
-from data_gradients.sample_iterable.detection import DetectionSampleIterable
+from data_gradients.dataset_adapters.detection_adapter import DetectionDatasetAdapter
+from data_gradients.sample_iterables.detection import DetectionSampleIterable
 
 
 class DetectionAnalysisManager(AnalysisManagerAbstract):
