+++ src/data_gradients/managers/segmentation_manager.py
@@ -4,9 +4,9 @@ import torch
 from data_gradients.config.utils import get_grouped_feature_extractors
 from data_gradients.managers.abstract_manager import AnalysisManagerAbstract
 from data_gradients.config.data.typing import SupportedDataType, FeatureExtractorsType
-from data_gradients.dataset_adapter.segmentation_adapter import SegmentationDatasetAdapter
+from data_gradients.dataset_adapters.segmentation_adapter import SegmentationDatasetAdapter
 from data_gradients.utils.summary_writer import SummaryWriter
-from data_gradients.sample_iterable.segmentation import SegmentationSampleIterable
+from data_gradients.sample_iterables.segmentation import SegmentationSampleIterable
 
 
 class SegmentationAnalysisManager(AnalysisManagerAbstract):
