+++ src/data_gradients/sample_iterables/contours.py
