+++ tests/unit_tests/batch_processor/test_detection_batch_formatter.py
@@ -1,6 +1,6 @@
 import torch
 import unittest
-from data_gradients.dataset_adapter.formatters.detection import DetectionBatchFormatter
+from data_gradients.dataset_adapters.formatters.detection import DetectionBatchFormatter
 from data_gradients.config.data import DetectionDataConfig
